app/requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements-dev.txt

# Include production requirements
-r requirements.txt

# Development and testing tools
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Development server with auto-reload
python-dotenv==1.0.0
